llamaindex-demo/pyproject.toml
[project]
name = "llamaindex-demo"
version = "0.1.0"
description = "LlamaIndex 企业级 LLM 应用开发框架演示项目"
authors = [
    { name = "maynard.liu", email = "415074476@qq.com" }
]
dependencies = [
    "llama-index>=0.14.8",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "pandas<2.3.0",
    "pillow>=12.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"
keywords = ["llm", "rag", "llamaindex", "ai", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llamaindex_demo"]
